+++ lib/src/scaffold.dart
@@ -8,15 +8,15 @@ import 'top_button.dart';
 class TopButton {
   /// `TopButton`'s title. Usually text that hints which action will be taken
   /// when user taps the button or pulls the page.
-  final String? title;
+  final String title;
 
   /// `TopButton`'s icon. Usually used for hinting which action will be taken
   /// when user taps the button or pulls the page.
-  final IconData? icon;
+  final IconData icon;
 
   /// The function that will be called when user pulls the page or taps the
   /// `TopButton`.
-  final Function? action;
+  final Function action;
 
   late final bool isHidden;
 
@@ -27,11 +27,11 @@ class TopButton {
   }) : isHidden = false;
 
   /// That function allows to hide the button.
-  TopButton.hidden()
+  /*TopButton.hidden()
       : isHidden = true,
         title = null,
         icon = null,
-        action = null;
+        action = null;*/
 }
 
 /// Widget that helps to build a page.
@@ -39,34 +39,43 @@ class TopButton {
 class SlikkerScaffold extends StatefulWidget {
   /// Text that is displayed above `header`. Useally is a text which
   /// indicates which page is it. In Material design it wuld be `AppBarTitle`
-  late final String title;
+  final String? title;
 
   /// Widget that is displayed above `content`. In Material Design it
   /// is usually `AppBar`.
-  late final Widget header;
+  final Widget? header;
 
-  /// Widget, usually `ListView`/`GridView` that contains other widgets.
-  /// Content of the page.
-  late final Widget content;
+  /// Content of the page. Place there buttons and main information for user.
+  final Widget? content;
 
   /// Constructor, which creates `TopButtonWidget`. The widget will be displayed
   /// in the top of the page. When user pulls the page or taps the widget,
   /// `TopButton.action` will be fired.
-  ///
-  /// If you want to **hide** the button, use `TopButton.hidden()`.
-  late final TopButton topButton;
+  // If you want to **hide** the button, use `TopButton.hidden()`.
+  final TopButton? topButton;
 
   /// Widget that is placed in the bottom of the screen, always visible, and
   /// floats above the `content`.
-  late final Widget floatingButton;
+  final Widget? floatingButton;
+
+  final bool sliver;
 
   SlikkerScaffold({
-    required this.topButton,
-    required this.title,
-    required this.header,
-    required this.content,
-    required this.floatingButton,
-  });
+    this.topButton,
+    this.title,
+    this.header,
+    this.content,
+    this.floatingButton,
+  }) : sliver = false;
+
+  /// Content of the page must be **Sliver** widget.
+  SlikkerScaffold.sliver({
+    this.topButton,
+    this.title,
+    this.header,
+    this.content,
+    this.floatingButton,
+  }) : sliver = true;
 
   @override
   _SlikkerScaffoldState createState() => _SlikkerScaffoldState();
@@ -98,11 +107,11 @@ class _SlikkerScaffoldState extends State<SlikkerScaffold> {
     if (info is ScrollUpdateNotification &&
         (percent != _scrollPreviousPercent || info.dragDetails == null) &&
         info.metrics.axisDirection == AxisDirection.down &&
-        !widget.topButton.isHidden) {
+        widget.topButton != null) {
       refreshTopButton(percent);
       _scrollPreviousPercent = percent;
       if (info.dragDetails == null && percent == 100 && !_scrollActionFired) {
-        widget.topButton.action!();
+        widget.topButton!.action();
         _scrollActionFired = true;
       }
       return true;
@@ -121,36 +130,53 @@ class _SlikkerScaffoldState extends State<SlikkerScaffold> {
           children: [
             NotificationListener<ScrollNotification>(
               onNotification: (scrollInfo) => scrolled(scrollInfo),
-              child: ListView(
+              child: CustomScrollView(
                 scrollDirection: Axis.vertical,
                 physics: AlwaysScrollableScrollPhysics(
-                    parent: BouncingScrollPhysics()),
-                children: <Widget>[
-                  Container(height: 52),
-                  if (!widget.topButton.isHidden)
-                    Center(
-                      child: TopButtonWidget(
-                        title: widget.topButton.title!,
-                        icon: widget.topButton.icon!,
-                        accent: 240,
-                        onTap: widget.topButton.action!,
-                        refresh: (Function topButtonFunction) =>
-                            refreshTopButton = topButtonFunction,
+                  parent: BouncingScrollPhysics(),
+                ),
+                slivers: <Widget>[
+                  SliverList(
+                    delegate: SliverChildListDelegate([
+                      Padding(
+                        padding: EdgeInsets.symmetric(vertical: 50),
+                        child: (widget.topButton == null)
+                            ? Container(height: 38)
+                            : Center(
+                                child: TopButtonWidget(
+                                  title: widget.topButton!.title,
+                                  icon: widget.topButton!.icon,
+                                  accent: 240,
+                                  onTap: widget.topButton!.action,
+                                  refresh: (Function topButtonFunction) =>
+                                      refreshTopButton = topButtonFunction,
+                                ),
+                              ),
                       ),
-                    ),
-                  Container(height: MediaQuery.of(context).size.height / 3.7),
-                  Text(
-                    widget.title,
-                    style: TextStyle(fontSize: 36.0),
-                    textAlign: TextAlign.center,
+                      Padding(
+                        padding: EdgeInsets.only(
+                          top: MediaQuery.of(context).size.height / 5,
+                        ),
+                      ),
+                      Text(
+                        widget.title ?? '',
+                        style: TextStyle(fontSize: 36.0),
+                        textAlign: TextAlign.center,
+                      ),
+                      Padding(
+                        padding: EdgeInsets.all(30),
+                        child: widget.header,
+                      ),
+                    ]),
                   ),
-                  Container(height: 20),
-                  widget.header,
-                  Padding(
-                    child: widget.content,
-                    padding: EdgeInsets.all(30),
+                  SliverPadding(
+                    sliver: widget.sliver
+                        ? widget.content
+                        : SliverToBoxAdapter(
+                            child: widget.content,
+                          ),
+                    padding: EdgeInsets.fromLTRB(30, 0, 30, 30),
                   ),
-                  Container(height: 60)
                 ],
               ),
             ),
@@ -158,10 +184,15 @@ class _SlikkerScaffoldState extends State<SlikkerScaffold> {
               alignment: Alignment.bottomCenter,
               child: Container(
                 decoration: BoxDecoration(
-                    gradient: LinearGradient(
-                        begin: Alignment.topCenter,
-                        end: Alignment(0, 0.25),
-                        colors: [Color(0x00F6F6FC), Color(0xFFF6F6FC)])),
+                  gradient: LinearGradient(
+                    begin: Alignment.topCenter,
+                    end: Alignment(0, 0.25),
+                    colors: [
+                      Color(0x00F6F6FC),
+                      Color(0xFFF6F6FC),
+                    ],
+                  ),
+                ),
                 height: 84,
               ),
             ),
@@ -169,7 +200,7 @@ class _SlikkerScaffoldState extends State<SlikkerScaffold> {
               alignment: Alignment.bottomCenter,
               child: Container(
                 child: widget.floatingButton,
-                margin: EdgeInsets.only(bottom: 25),
+                margin: EdgeInsets.only(bottom: 30),
               ),
             ),
           ],